+++ apps/dashboard/src/pages/environments.tsx
@@ -1,37 +1,23 @@
 import { PageMeta } from '@/components/page-meta';
 import { ApiServiceLevelEnum } from '@novu/shared';
 import { useEffect } from 'react';
-import { RiBookMarkedLine, RiSparkling2Line } from 'react-icons/ri';
-import { useNavigate } from 'react-router-dom';
 import { CreateEnvironmentButton } from '../components/create-environment-button';
 import { DashboardLayout } from '../components/dashboard-layout';
 import { EnvironmentsList } from '../components/environments-list';
-import { Badge } from '../components/primitives/badge';
-import { Button } from '../components/primitives/button';
-import { LinkButton } from '../components/primitives/button-link';
-import { CopyButton } from '../components/primitives/copy-button';
-import { EnvironmentBranchIcon } from '../components/primitives/environment-branch-icon';
-import { Separator } from '../components/primitives/separator';
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '../components/primitives/table';
-import TruncatedText from '../components/truncated-text';
-import { useAuth } from '../context/auth/hooks';
-import { useEnvironment, useFetchEnvironments } from '../context/environment/hooks';
-import { useCheckoutSession } from '../hooks/use-checkout-session';
+import { FreeTierEmptyState } from '../components/free-tier-empty-state';
 import { useFetchSubscription } from '../hooks/use-fetch-subscription';
 import { useTelemetry } from '../hooks/use-telemetry';
-import { ROUTES } from '../utils/routes';
 import { TelemetryEvent } from '../utils/telemetry';
 
 export function EnvironmentsPage() {
   const track = useTelemetry();
-  const { currentEnvironment } = useEnvironment();
   const { subscription } = useFetchSubscription();
 
   const isPaidTier =
     subscription?.apiServiceLevel === ApiServiceLevelEnum.BUSINESS ||
     subscription?.apiServiceLevel === ApiServiceLevelEnum.ENTERPRISE;
   const isTrialActive = subscription?.trial?.isActive;
-  const canAccessEnvironments = isPaidTier && !isTrialActive;
+  const canAccessEnvironments = !isPaidTier && !isTrialActive;
 
   useEffect(() => {
     track(TelemetryEvent.ENVIRONMENTS_PAGE_VIEWED);
@@ -41,7 +27,7 @@ export function EnvironmentsPage() {
     <>
       <PageMeta title={`Environments`} />
       <DashboardLayout headerStartItems={<h1 className="text-foreground-950">Environments</h1>}>
-        {!track ? (
+        {canAccessEnvironments ? (
           <div className="flex flex-col justify-between gap-2 px-2.5 py-2.5">
             <div className="flex justify-end">
               <CreateEnvironmentButton />
@@ -53,317 +39,5 @@ export function EnvironmentsPage() {
         )}
       </DashboardLayout>
     </>
-  );
-}
-
-function FreeTierEmptyState() {
-  const navigate = useNavigate();
-  const { navigateToCheckout } = useCheckoutSession();
-  const { currentOrganization } = useAuth();
-  const { environments = [] } = useFetchEnvironments({
-    organizationId: currentOrganization?._id,
-  });
-  const { currentEnvironment } = useEnvironment();
-
-  return (
-    <div className="flex h-full w-full flex-col items-center justify-center gap-6 px-4">
-      <div className="flex w-full max-w-[480px] flex-col items-center gap-6 text-center">
-        <div className="flex w-full flex-col gap-3">
-          <div className="flex flex-col items-center gap-2">
-            <div className="mb-[50px]">
-              <EmptyStateSvg />
-            </div>
-            <h2 className="text-foreground-900 text-label-md">Manage Your Environments</h2>
-            <p className="text-text-soft text-label-xs mb-3 max-w-[300px]">
-              Create additional environments to test, stage, or experiment without affecting your live systems.
-            </p>
-          </div>
-          <Separator variant="line-text">YOUR ENVIRONMENTS</Separator>
-          <div className="flex w-full items-center justify-center px-5">
-            <Table>
-              <TableHeader className="w-full">
-                <TableRow>
-                  <TableHead>Name</TableHead>
-                  <TableHead>Identifier</TableHead>
-                </TableRow>
-              </TableHeader>
-              <TableBody>
-                {environments.map((environment) => (
-                  <TableRow key={environment._id} className="group relative isolate">
-                    <TableCell className="font-medium">
-                      <div className="flex items-center gap-2">
-                        <EnvironmentBranchIcon size="sm" environment={environment} />
-                        <div className="flex items-center gap-1">
-                          <TruncatedText className="max-w-[32ch]">{environment.name}</TruncatedText>
-                          {environment._id === currentEnvironment?._id && (
-                            <Badge color="blue" size="sm" variant="lighter">
-                              Current
-                            </Badge>
-                          )}
-                        </div>
-                      </div>
-                    </TableCell>
-                    <TableCell>
-                      <div className="flex items-center gap-1 transition-opacity duration-200">
-                        <TruncatedText className="text-foreground-400 font-code block text-xs">
-                          {environment.identifier}
-                        </TruncatedText>
-                        <CopyButton
-                          className="z-10 flex size-2 p-0 px-1 opacity-0 group-hover:opacity-100"
-                          valueToCopy={environment.identifier}
-                          size="2xs"
-                          mode="ghost"
-                        />
-                      </div>
-                    </TableCell>
-                  </TableRow>
-                ))}
-              </TableBody>
-            </Table>
-          </div>
-        </div>
-
-        <div>
-          <Button
-            variant="primary"
-            mode="gradient"
-            size="xs"
-            className="mb-2"
-            onClick={() => navigateToCheckout('month')}
-            leadingIcon={RiSparkling2Line}
-          >
-            Upgrade to Business Tier
-          </Button>
-          <LinkButton size="sm" leadingIcon={RiBookMarkedLine} onClick={() => navigate(ROUTES.SETTINGS_BILLING)}>
-            How does this help?
-          </LinkButton>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function EmptyStateSvg() {
-  return (
-    <svg width="325" height="112" viewBox="0 0 325 112" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <path
-        d="M90 93.5H67.794C67.2636 93.5 66.7549 93.2893 66.3798 92.9142C66.0047 92.5391 65.794 92.0304 65.794 91.5V59.5C65.794 58.9696 65.5833 58.4609 65.2082 58.0858C64.8331 57.7107 64.3244 57.5 63.794 57.5H47"
-        stroke="#BCC3CE"
-        stroke-width="0.5"
-        stroke-linejoin="round"
-        stroke-dasharray="4 4"
-      />
-      <path
-        d="M90 93.5H67.794C67.2636 93.5 66.7549 93.2893 66.3798 92.9142C66.0047 92.5391 65.794 92.0304 65.794 91.5V59.5C65.794 58.9696 65.5833 58.4609 65.2082 58.0858C64.8331 57.7107 64.3244 57.5 63.794 57.5H47"
-        stroke="url(#paint0_linear_12184_111598)"
-        stroke-width="0.5"
-        stroke-linejoin="round"
-        stroke-dasharray="4 4"
-      />
-      <path
-        d="M235 93.5H262.207C262.737 93.5 263.246 93.2893 263.621 92.9142C263.996 92.5391 264.207 92.0304 264.207 91.5V59.5C264.207 58.9696 264.418 58.4609 264.793 58.0858C265.168 57.7107 265.677 57.5 266.207 57.5"
-        stroke="url(#paint1_linear_12184_111598)"
-        stroke-width="0.5"
-        stroke-linejoin="round"
-        stroke-dasharray="4 4"
-      />
-      <path
-        d="M90 21.5H67.794C67.2636 21.5 66.7549 21.7107 66.3798 22.0858C66.0047 22.4609 65.794 22.9696 65.794 23.5V55.5C65.794 56.0304 65.5833 56.5391 65.2082 56.9142C64.8331 57.2893 64.3244 57.5 63.794 57.5H47"
-        stroke="#DD2450"
-        stroke-width="0.5"
-        stroke-linejoin="round"
-        stroke-dasharray="4 4"
-      />
-      <path
-        d="M278 57.5H266.794C266.264 57.5 265.755 57.2893 265.38 56.9142C265.005 56.5391 264.794 56.0304 264.794 55.5V23.5C264.794 22.9696 264.583 22.4609 264.208 22.0858C263.833 21.7107 263.324 21.5 262.794 21.5H235"
-        stroke="#DD2450"
-        stroke-width="0.5"
-        stroke-linejoin="round"
-        stroke-dasharray="4 4"
-      />
-      <rect x="283.375" y="36.875" width="41.25" height="41.25" rx="7.625" stroke="#DD2450" stroke-width="0.75" />
-      <rect x="287" y="40.5" width="34" height="34" rx="6" fill="white" />
-      <rect
-        x="287.375"
-        y="40.875"
-        width="33.25"
-        height="33.25"
-        rx="5.625"
-        stroke="#FB3748"
-        stroke-opacity="0.24"
-        stroke-width="0.75"
-      />
-      <path
-        d="M303.359 53.3991L303.37 53.3969V53.3857V53C303.37 52.652 303.652 52.3708 304 52.3708C304.348 52.3708 304.629 52.652 304.629 53V53.3857V53.3969L304.64 53.3991C306.1 53.6952 307.2 54.9875 307.2 56.5357V56.9134C307.2 57.863 307.549 58.7763 308.178 59.4859L308.178 59.4859L308.327 59.6527L308.327 59.6527C308.492 59.8375 308.533 60.1029 308.431 60.329C308.329 60.5552 308.104 60.7006 307.857 60.7006H300.142C299.895 60.7006 299.669 60.5551 299.568 60.3291C299.468 60.1029 299.507 59.8374 299.672 59.6527L299.672 59.6527L299.821 59.4859L299.821 59.4859C300.45 58.7763 300.799 57.861 300.799 56.9134V56.5357C300.799 54.9875 301.899 53.6952 303.359 53.3991ZM305.272 61.3709C305.268 61.7038 305.135 62.0223 304.9 62.2576C304.661 62.496 304.337 62.6292 304 62.6292C303.662 62.6292 303.338 62.496 303.099 62.2576C302.864 62.0223 302.731 61.7038 302.728 61.3709H304H305.272Z"
-        fill="#DD2450"
-        stroke="#DD2450"
-        stroke-width="0.0273438"
-      />
-      <rect x="0.375" y="36.875" width="41.25" height="41.25" rx="7.625" stroke="#DD2450" stroke-width="0.75" />
-      <rect x="4" y="40.5" width="34" height="34" rx="6" fill="white" />
-      <rect
-        x="4.375"
-        y="40.875"
-        width="33.25"
-        height="33.25"
-        rx="5.625"
-        stroke="#FB3748"
-        stroke-opacity="0.24"
-        stroke-width="0.75"
-      />
-      <path
-        d="M11 57.5C11 51.9772 15.4772 47.5 21 47.5C26.5228 47.5 31 51.9772 31 57.5C31 63.0228 26.5228 67.5 21 67.5C15.4772 67.5 11 63.0228 11 57.5Z"
-        fill="white"
-      />
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M24.24 56.3098C24.24 56.6323 23.8485 56.792 23.6227 56.5614L19.0035 51.8401C19.6449 51.6144 20.32 51.4994 21 51.5C22.1936 51.5 23.3055 51.8488 24.24 52.4491V56.3098ZM25.92 54.065V56.3098C25.92 58.1379 23.7004 59.0431 22.422 57.7363L17.4544 52.6591C15.966 53.7511 15 55.5129 15 57.5C15 58.7776 15.3994 59.9619 16.08 60.935V58.7023C16.08 56.8741 18.2996 55.9689 19.578 57.2758L24.5389 62.3458C26.031 61.2545 27 59.4905 27 57.5C27 56.2224 26.6006 55.0381 25.92 54.065ZM18.3772 58.4506L22.9879 63.1625C22.3658 63.3811 21.6968 63.5 21 63.5C19.8067 63.5 18.6945 63.1513 17.76 62.5509V58.7023C17.76 58.3798 18.1519 58.22 18.3772 58.4506Z"
-        fill="url(#paint2_radial_12184_111598)"
-      />
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M24.24 56.3098C24.24 56.6323 23.8485 56.792 23.6227 56.5614L19.0035 51.8401C19.6449 51.6144 20.32 51.4994 21 51.5C22.1936 51.5 23.3055 51.8488 24.24 52.4491V56.3098ZM25.92 54.065V56.3098C25.92 58.1379 23.7004 59.0431 22.422 57.7363L17.4544 52.6591C15.966 53.7511 15 55.5129 15 57.5C15 58.7776 15.3994 59.9619 16.08 60.935V58.7023C16.08 56.8741 18.2996 55.9689 19.578 57.2758L24.5389 62.3458C26.031 61.2545 27 59.4905 27 57.5C27 56.2224 26.6006 55.0381 25.92 54.065ZM18.3772 58.4506L22.9879 63.1625C22.3658 63.3811 21.6968 63.5 21 63.5C19.8067 63.5 18.6945 63.1513 17.76 62.5509V58.7023C17.76 58.3798 18.1519 58.22 18.3772 58.4506Z"
-        fill="url(#paint3_linear_12184_111598)"
-      />
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M24.24 56.3098C24.24 56.6323 23.8485 56.792 23.6227 56.5614L19.0035 51.8401C19.6449 51.6144 20.32 51.4994 21 51.5C22.1936 51.5 23.3055 51.8488 24.24 52.4491V56.3098ZM25.92 54.065V56.3098C25.92 58.1379 23.7004 59.0431 22.422 57.7363L17.4544 52.6591C15.966 53.7511 15 55.5129 15 57.5C15 58.7776 15.3994 59.9619 16.08 60.935V58.7023C16.08 56.8741 18.2996 55.9689 19.578 57.2758L24.5389 62.3458C26.031 61.2545 27 59.4905 27 57.5C27 56.2224 26.6006 55.0381 25.92 54.065ZM18.3772 58.4506L22.9879 63.1625C22.3658 63.3811 21.6968 63.5 21 63.5C19.8067 63.5 18.6945 63.1513 17.76 62.5509V58.7023C17.76 58.3798 18.1519 58.22 18.3772 58.4506Z"
-        fill="url(#paint4_linear_12184_111598)"
-      />
-      <rect x="94.375" y="0.875" width="60.7672" height="41.25" rx="7.625" stroke="#E1E4EA" stroke-width="0.75" />
-      <rect x="98.375" y="4.875" width="52.7672" height="33.25" rx="5.625" fill="white" />
-      <rect x="98.375" y="4.875" width="52.7672" height="33.25" rx="5.625" stroke="#F2F5F8" stroke-width="0.75" />
-      <path
-        d="M112.973 21.6875C112.973 21.8049 113.189 22.0092 113.691 22.2099C114.341 22.4694 115.262 22.625 116.259 22.625C117.255 22.625 118.177 22.4694 118.826 22.2099C119.328 22.0092 119.544 21.8049 119.544 21.6875V20.8734C118.77 21.2559 117.586 21.5 116.259 21.5C114.932 21.5 113.747 21.2555 112.973 20.8734V21.6875ZM119.544 22.7484C118.77 23.1309 117.586 23.375 116.259 23.375C114.932 23.375 113.747 23.1305 112.973 22.7484V23.5625C112.973 23.6799 113.189 23.8842 113.691 24.0849C114.341 24.3444 115.262 24.5 116.259 24.5C117.255 24.5 118.177 24.3444 118.826 24.0849C119.328 23.8842 119.544 23.6799 119.544 23.5625V22.7484ZM112.034 23.5625V19.8125C112.034 18.8806 113.926 18.125 116.259 18.125C118.592 18.125 120.483 18.8806 120.483 19.8125V23.5625C120.483 24.4944 118.592 25.25 116.259 25.25C113.926 25.25 112.034 24.4944 112.034 23.5625ZM116.259 20.75C117.255 20.75 118.177 20.5944 118.826 20.3349C119.328 20.1342 119.544 19.9299 119.544 19.8125C119.544 19.6951 119.328 19.4907 118.826 19.2901C118.177 19.0306 117.255 18.875 116.259 18.875C115.262 18.875 114.341 19.0306 113.691 19.2901C113.189 19.4907 112.973 19.6951 112.973 19.8125C112.973 19.9299 113.189 20.1342 113.691 20.3349C114.341 20.5944 115.262 20.75 116.259 20.75Z"
-        fill="#FF8447"
-      />
-      <path
-        d="M125.486 24.5V18.66H127.118C127.502 18.66 127.835 18.7347 128.118 18.884C128.401 19.028 128.619 19.2333 128.774 19.5C128.934 19.7667 129.014 20.0813 129.014 20.444V22.708C129.014 23.0653 128.934 23.38 128.774 23.652C128.619 23.924 128.401 24.1347 128.118 24.284C127.835 24.428 127.502 24.5 127.118 24.5H125.486ZM126.35 23.716H127.118C127.438 23.716 127.689 23.628 127.87 23.452C128.057 23.2707 128.15 23.0227 128.15 22.708V20.444C128.15 20.1347 128.057 19.892 127.87 19.716C127.689 19.5347 127.438 19.444 127.118 19.444H126.35V23.716ZM130.371 24.5V18.66H133.803V19.428H131.219V21.084H133.523V21.844H131.219V23.732H133.803V24.5H130.371ZM136.24 24.5L134.76 18.66H135.648L136.584 22.548C136.642 22.7773 136.693 23.0067 136.736 23.236C136.784 23.46 136.818 23.636 136.84 23.764C136.861 23.636 136.89 23.46 136.928 23.236C136.97 23.0067 137.021 22.7747 137.08 22.54L138.008 18.66H138.872L137.384 24.5H136.24Z"
-        fill="#FF8447"
-      />
-      <rect x="164.892" y="0.875" width="65.7672" height="41.25" rx="7.625" stroke="#E1E4EA" stroke-width="0.75" />
-      <rect x="168.892" y="4.875" width="57.7672" height="33.25" rx="5.625" fill="white" />
-      <rect x="168.892" y="4.875" width="57.7672" height="33.25" rx="5.625" stroke="#F2F5F8" stroke-width="0.75" />
-      <path
-        d="M183.49 21.6875C183.49 21.8049 183.706 22.0092 184.208 22.2099C184.858 22.4694 185.779 22.625 186.776 22.625C187.772 22.625 188.694 22.4694 189.343 22.2099C189.845 22.0092 190.062 21.8049 190.062 21.6875V20.8734C189.287 21.2559 188.103 21.5 186.776 21.5C185.449 21.5 184.265 21.2555 183.49 20.8734V21.6875ZM190.062 22.7484C189.287 23.1309 188.103 23.375 186.776 23.375C185.449 23.375 184.265 23.1305 183.49 22.7484V23.5625C183.49 23.6799 183.706 23.8842 184.208 24.0849C184.858 24.3444 185.779 24.5 186.776 24.5C187.772 24.5 188.694 24.3444 189.343 24.0849C189.845 23.8842 190.062 23.6799 190.062 23.5625V22.7484ZM182.551 23.5625V19.8125C182.551 18.8806 184.443 18.125 186.776 18.125C189.109 18.125 191 18.8806 191 19.8125V23.5625C191 24.4944 189.109 25.25 186.776 25.25C184.443 25.25 182.551 24.4944 182.551 23.5625ZM186.776 20.75C187.772 20.75 188.694 20.5944 189.343 20.3349C189.845 20.1342 190.062 19.9299 190.062 19.8125C190.062 19.6951 189.845 19.4907 189.343 19.2901C188.694 19.0306 187.772 18.875 186.776 18.875C185.779 18.875 184.858 19.0306 184.208 19.2901C183.706 19.4907 183.49 19.6951 183.49 19.8125C183.49 19.9299 183.706 20.1342 184.208 20.3349C184.858 20.5944 185.779 20.75 186.776 20.75Z"
-        fill="#7D52F4"
-      />
-      <path
-        d="M196.113 24.5V18.66H198.025C198.403 18.66 198.731 18.732 199.009 18.876C199.291 19.02 199.507 19.2253 199.657 19.492C199.811 19.7533 199.889 20.0627 199.889 20.42C199.889 20.772 199.811 21.0813 199.657 21.348C199.502 21.6147 199.286 21.82 199.009 21.964C198.731 22.108 198.403 22.18 198.025 22.18H196.977V24.5H196.113ZM196.977 21.404H198.025C198.323 21.404 198.561 21.316 198.737 21.14C198.918 20.9587 199.009 20.7187 199.009 20.42C199.009 20.116 198.918 19.876 198.737 19.7C198.561 19.524 198.323 19.436 198.025 19.436H196.977V21.404ZM200.91 24.5V18.66H202.742C203.11 18.66 203.43 18.732 203.702 18.876C203.974 19.0147 204.184 19.212 204.334 19.468C204.488 19.724 204.566 20.0253 204.566 20.372C204.566 20.7613 204.464 21.1 204.262 21.388C204.064 21.6707 203.795 21.8707 203.454 21.988L204.646 24.5H203.662L202.598 22.1H201.774V24.5H200.91ZM201.774 21.34H202.742C203.03 21.34 203.259 21.2547 203.43 21.084C203.6 20.908 203.686 20.676 203.686 20.388C203.686 20.0893 203.6 19.8547 203.43 19.684C203.259 19.5133 203.03 19.428 202.742 19.428H201.774V21.34ZM207.434 24.58C207.072 24.58 206.757 24.5133 206.49 24.38C206.229 24.2413 206.026 24.044 205.882 23.788C205.744 23.5267 205.674 23.2227 205.674 22.876V20.284C205.674 19.932 205.744 19.628 205.882 19.372C206.026 19.116 206.229 18.9213 206.49 18.788C206.757 18.6493 207.072 18.58 207.434 18.58C207.797 18.58 208.109 18.6493 208.37 18.788C208.637 18.9213 208.84 19.116 208.978 19.372C209.122 19.628 209.194 19.9293 209.194 20.276V22.876C209.194 23.2227 209.122 23.5267 208.978 23.788C208.84 24.044 208.637 24.2413 208.37 24.38C208.109 24.5133 207.797 24.58 207.434 24.58ZM207.434 23.812C207.728 23.812 207.949 23.732 208.098 23.572C208.253 23.4067 208.33 23.1747 208.33 22.876V20.284C208.33 19.98 208.253 19.748 208.098 19.588C207.949 19.428 207.728 19.348 207.434 19.348C207.146 19.348 206.925 19.428 206.77 19.588C206.616 19.748 206.538 19.98 206.538 20.284V22.876C206.538 23.1747 206.616 23.4067 206.77 23.572C206.925 23.732 207.146 23.812 207.434 23.812ZM210.495 24.5V18.66H212.127C212.511 18.66 212.845 18.7347 213.127 18.884C213.41 19.028 213.629 19.2333 213.783 19.5C213.943 19.7667 214.023 20.0813 214.023 20.444V22.708C214.023 23.0653 213.943 23.38 213.783 23.652C213.629 23.924 213.41 24.1347 213.127 24.284C212.845 24.428 212.511 24.5 212.127 24.5H210.495ZM211.359 23.716H212.127C212.447 23.716 212.698 23.628 212.879 23.452C213.066 23.2707 213.159 23.0227 213.159 22.708V20.444C213.159 20.1347 213.066 19.892 212.879 19.716C212.698 19.5347 212.447 19.444 212.127 19.444H211.359V23.716Z"
-        fill="#7D52F4"
-      />
-      <rect x="90.375" y="69.875" width="144.305" height="41.25" rx="7.625" stroke="#E1E4EA" stroke-width="0.75" />
-      <rect x="94.375" y="73.875" width="136.305" height="33.25" rx="5.625" fill="white" />
-      <rect x="94.375" y="73.875" width="136.305" height="33.25" rx="5.625" stroke="#F2F5F8" stroke-width="0.75" />
-      <path
-        d="M108.973 90.6875C108.973 90.8049 109.189 91.0092 109.691 91.2099C110.341 91.4694 111.262 91.625 112.259 91.625C113.255 91.625 114.177 91.4694 114.826 91.2099C115.328 91.0092 115.544 90.8049 115.544 90.6875V89.8734C114.77 90.2559 113.586 90.5 112.259 90.5C110.932 90.5 109.747 90.2555 108.973 89.8734V90.6875ZM115.544 91.7484C114.77 92.1309 113.586 92.375 112.259 92.375C110.932 92.375 109.747 92.1305 108.973 91.7484V92.5625C108.973 92.6799 109.189 92.8842 109.691 93.0849C110.341 93.3444 111.262 93.5 112.259 93.5C113.255 93.5 114.177 93.3444 114.826 93.0849C115.328 92.8842 115.544 92.6799 115.544 92.5625V91.7484ZM108.034 92.5625V88.8125C108.034 87.8806 109.926 87.125 112.259 87.125C114.592 87.125 116.483 87.8806 116.483 88.8125V92.5625C116.483 93.4944 114.592 94.25 112.259 94.25C109.926 94.25 108.034 93.4944 108.034 92.5625ZM112.259 89.75C113.255 89.75 114.177 89.5944 114.826 89.3349C115.328 89.1342 115.544 88.9299 115.544 88.8125C115.544 88.6951 115.328 88.4907 114.826 88.2901C114.177 88.0306 113.255 87.875 112.259 87.875C111.262 87.875 110.341 88.0306 109.691 88.2901C109.189 88.4907 108.973 88.6951 108.973 88.8125C108.973 88.9299 109.189 89.1342 109.691 89.3349C110.341 89.5944 111.262 89.75 112.259 89.75Z"
-        fill="#CACFD8"
-      />
-      <path
-        d="M121.675 93.5V87.66H125.107V88.428H122.523V90.084H124.827V90.844H122.523V92.732H125.107V93.5H121.675ZM125.992 93.5L127.64 90.524L126.088 87.66H127.064L127.88 89.252C127.928 89.3533 127.976 89.4547 128.024 89.556C128.072 89.6573 128.11 89.7347 128.136 89.788C128.158 89.7347 128.192 89.6573 128.24 89.556C128.288 89.4547 128.336 89.3533 128.384 89.252L129.216 87.66H130.152L128.6 90.492L130.248 93.5H129.272L128.368 91.772C128.315 91.6707 128.264 91.5693 128.216 91.468C128.168 91.3613 128.131 91.2787 128.104 91.22C128.083 91.2787 128.048 91.3587 128 91.46C127.952 91.5613 127.904 91.6627 127.856 91.764L126.936 93.5H125.992ZM131.189 93.5V87.66H133.101C133.48 87.66 133.808 87.732 134.085 87.876C134.368 88.02 134.584 88.2253 134.733 88.492C134.888 88.7533 134.965 89.0627 134.965 89.42C134.965 89.772 134.888 90.0813 134.733 90.348C134.579 90.6147 134.363 90.82 134.085 90.964C133.808 91.108 133.48 91.18 133.101 91.18H132.053V93.5H131.189ZM132.053 90.404H133.101C133.4 90.404 133.637 90.316 133.813 90.14C133.995 89.9587 134.085 89.7187 134.085 89.42C134.085 89.116 133.995 88.876 133.813 88.7C133.637 88.524 133.4 88.436 133.101 88.436H132.053V90.404ZM136.01 93.5V92.732H137.562V88.412L136.002 89.572V88.636L137.338 87.66H138.426V92.732H139.682V93.5H136.01Z"
-        fill="#CACFD8"
-      />
-      <path
-        d="M152.085 90.725C152.085 90.8658 152.344 91.1111 152.946 91.3518C153.726 91.6632 154.832 91.85 156.028 91.85C157.223 91.85 158.329 91.6632 159.109 91.3518C159.711 91.1111 159.971 90.8658 159.971 90.725V89.748C159.041 90.207 157.62 90.5 156.028 90.5C154.435 90.5 153.014 90.2066 152.085 89.748V90.725ZM159.971 91.998C159.041 92.457 157.62 92.75 156.028 92.75C154.435 92.75 153.014 92.4566 152.085 91.998V92.975C152.085 93.1158 152.344 93.3611 152.946 93.6018C153.726 93.9132 154.832 94.1 156.028 94.1C157.223 94.1 158.329 93.9132 159.109 93.6018C159.711 93.3611 159.971 93.1158 159.971 92.975V91.998ZM150.958 92.975V88.475C150.958 87.3567 153.228 86.45 156.028 86.45C158.827 86.45 161.097 87.3567 161.097 88.475V92.975C161.097 94.0932 158.827 95 156.028 95C153.228 95 150.958 94.0932 150.958 92.975ZM156.028 89.6C157.223 89.6 158.329 89.4132 159.109 89.1018C159.711 88.8611 159.971 88.6158 159.971 88.475C159.971 88.3341 159.711 88.0889 159.109 87.8481C158.329 87.5367 157.223 87.35 156.028 87.35C154.832 87.35 153.726 87.5367 152.946 87.8481C152.344 88.0889 152.085 88.3341 152.085 88.475C152.085 88.6158 152.344 88.8611 152.946 89.1018C153.726 89.4132 154.832 89.6 156.028 89.6Z"
-        fill="#CACFD8"
-      />
-      <path
-        d="M169.376 95.12L168.404 93.509L168.548 93.581C168.524 93.581 168.494 93.581 168.458 93.581C168.422 93.587 168.383 93.59 168.341 93.59C167.933 93.59 167.576 93.509 167.27 93.347C166.964 93.185 166.727 92.96 166.559 92.672C166.391 92.378 166.307 92.036 166.307 91.646V88.784C166.307 88.388 166.391 88.046 166.559 87.758C166.727 87.47 166.964 87.245 167.27 87.083C167.576 86.921 167.933 86.84 168.341 86.84C168.755 86.84 169.112 86.921 169.412 87.083C169.718 87.245 169.955 87.47 170.123 87.758C170.291 88.046 170.375 88.388 170.375 88.784V91.646C170.375 92.042 170.288 92.39 170.114 92.69C169.946 92.99 169.703 93.215 169.385 93.365L170.456 95.12H169.376ZM168.341 92.735C168.671 92.735 168.929 92.636 169.115 92.438C169.307 92.24 169.403 91.976 169.403 91.646V88.784C169.403 88.448 169.307 88.184 169.115 87.992C168.929 87.794 168.671 87.695 168.341 87.695C168.017 87.695 167.759 87.794 167.567 87.992C167.375 88.184 167.279 88.448 167.279 88.784V91.646C167.279 91.976 167.372 92.24 167.558 92.438C167.75 92.636 168.011 92.735 168.341 92.735ZM171.425 93.5L173.099 86.93H174.377L176.051 93.5H175.07L174.674 91.835H172.811L172.415 93.5H171.425ZM172.991 91.034H174.485L174.035 89.135C173.957 88.799 173.891 88.508 173.837 88.262C173.789 88.01 173.756 87.839 173.738 87.749C173.72 87.839 173.687 88.01 173.639 88.262C173.591 88.508 173.525 88.796 173.441 89.126L172.991 91.034Z"
-        fill="#CACFD8"
-      />
-      <path
-        d="M187.511 90.6875C187.511 90.8049 187.727 91.0092 188.229 91.2099C188.879 91.4694 189.8 91.625 190.797 91.625C191.793 91.625 192.715 91.4694 193.364 91.2099C193.866 91.0092 194.082 90.8049 194.082 90.6875V89.8734C193.308 90.2559 192.124 90.5 190.797 90.5C189.47 90.5 188.285 90.2555 187.511 89.8734V90.6875ZM194.082 91.7484C193.308 92.1309 192.124 92.375 190.797 92.375C189.47 92.375 188.285 92.1305 187.511 91.7484V92.5625C187.511 92.6799 187.727 92.8842 188.229 93.0849C188.879 93.3444 189.8 93.5 190.797 93.5C191.793 93.5 192.715 93.3444 193.364 93.0849C193.866 92.8842 194.082 92.6799 194.082 92.5625V91.7484ZM186.572 92.5625V88.8125C186.572 87.8806 188.464 87.125 190.797 87.125C193.129 87.125 195.021 87.8806 195.021 88.8125V92.5625C195.021 93.4944 193.129 94.25 190.797 94.25C188.464 94.25 186.572 93.4944 186.572 92.5625ZM190.797 89.75C191.793 89.75 192.715 89.5944 193.364 89.3349C193.866 89.1342 194.082 88.9299 194.082 88.8125C194.082 88.6951 193.866 88.4907 193.364 88.2901C192.715 88.0306 191.793 87.875 190.797 87.875C189.8 87.875 188.879 88.0306 188.229 88.2901C187.727 88.4907 187.511 88.6951 187.511 88.8125C187.511 88.9299 187.727 89.1342 188.229 89.3349C188.879 89.5944 189.8 89.75 190.797 89.75Z"
-        fill="#CACFD8"
-      />
-      <path
-        d="M201.885 93.58C201.491 93.58 201.152 93.516 200.869 93.388C200.587 93.2547 200.368 93.068 200.213 92.828C200.059 92.5827 199.981 92.292 199.981 91.956H200.837C200.837 92.2227 200.931 92.4333 201.117 92.588C201.304 92.7373 201.563 92.812 201.893 92.812C202.203 92.812 202.445 92.7373 202.621 92.588C202.797 92.4387 202.885 92.2333 202.885 91.972C202.885 91.7533 202.824 91.564 202.701 91.404C202.584 91.244 202.413 91.1347 202.189 91.076L201.453 90.86C201.027 90.7373 200.696 90.5293 200.461 90.236C200.232 89.9373 200.117 89.5827 200.117 89.172C200.117 88.852 200.189 88.572 200.333 88.332C200.477 88.092 200.683 87.9053 200.949 87.772C201.216 87.6387 201.531 87.572 201.893 87.572C202.427 87.572 202.853 87.716 203.173 88.004C203.499 88.292 203.664 88.6787 203.669 89.164H202.805C202.805 88.908 202.723 88.708 202.557 88.564C202.397 88.4147 202.171 88.34 201.877 88.34C201.595 88.34 201.373 88.4093 201.213 88.548C201.053 88.6813 200.973 88.8707 200.973 89.116C200.973 89.3347 201.032 89.524 201.149 89.684C201.272 89.844 201.445 89.956 201.669 90.02L202.413 90.244C202.84 90.3613 203.168 90.5693 203.397 90.868C203.627 91.1667 203.741 91.524 203.741 91.94C203.741 92.2653 203.664 92.5533 203.509 92.804C203.355 93.0493 203.139 93.2413 202.861 93.38C202.584 93.5133 202.259 93.58 201.885 93.58ZM206.226 93.5V88.452H204.666V87.652H208.65V88.452H207.09V93.5H206.226ZM209.399 93.5L210.887 87.66H212.023L213.511 93.5H212.639L212.287 92.02H210.631L210.279 93.5H209.399ZM210.791 91.308H212.119L211.719 89.62C211.65 89.3213 211.591 89.0627 211.543 88.844C211.501 88.62 211.471 88.468 211.455 88.388C211.439 88.468 211.41 88.62 211.367 88.844C211.325 89.0627 211.266 89.3187 211.191 89.612L210.791 91.308ZM216.284 93.58C215.921 93.58 215.604 93.5133 215.332 93.38C215.065 93.2413 214.857 93.044 214.708 92.788C214.564 92.5267 214.492 92.2227 214.492 91.876V89.284C214.492 88.932 214.564 88.628 214.708 88.372C214.857 88.116 215.065 87.9213 215.332 87.788C215.604 87.6493 215.921 87.58 216.284 87.58C216.647 87.58 216.961 87.6493 217.228 87.788C217.495 87.9267 217.7 88.124 217.844 88.38C217.993 88.6307 218.068 88.932 218.068 89.284H217.204C217.204 88.98 217.124 88.748 216.964 88.588C216.804 88.428 216.577 88.348 216.284 88.348C215.991 88.348 215.761 88.428 215.596 88.588C215.436 88.748 215.356 88.9773 215.356 89.276V91.876C215.356 92.1747 215.436 92.4067 215.596 92.572C215.761 92.7373 215.991 92.82 216.284 92.82C216.577 92.82 216.804 92.7373 216.964 92.572C217.124 92.4067 217.204 92.1747 217.204 91.876V91.22H216.116V90.46H218.068V91.876C218.068 92.2227 217.993 92.524 217.844 92.78C217.7 93.036 217.495 93.2333 217.228 93.372C216.961 93.5107 216.647 93.58 216.284 93.58Z"
-        fill="#CACFD8"
-      />
-      <rect x="95" y="78.5" width="134" height="25" fill="url(#paint5_linear_12184_111598)" />
-      <rect x="95" y="78.5" width="134" height="25" fill="url(#paint6_linear_12184_111598)" />
-      <defs>
-        <linearGradient
-          id="paint0_linear_12184_111598"
-          x1="50.5903"
-          y1="84.5"
-          x2="79.4364"
-          y2="110.516"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#BCC3CE" />
-          <stop offset="0.759383" stop-color="white" stop-opacity="0.5" />
-        </linearGradient>
-        <linearGradient
-          id="paint1_linear_12184_111598"
-          x1="266.001"
-          y1="82.5"
-          x2="220"
-          y2="94"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#BCC3CE" />
-          <stop offset="1" stop-color="white" stop-opacity="0.5" />
-        </linearGradient>
-        <radialGradient
-          id="paint2_radial_12184_111598"
-          cx="0"
-          cy="0"
-          r="1"
-          gradientUnits="userSpaceOnUse"
-          gradientTransform="translate(20.9999 57.5003) rotate(135) scale(8.48527)"
-        >
-          <stop offset="0.34" stop-color="#FF006A" />
-          <stop offset="0.613" stop-color="#E300BD" />
-          <stop offset="0.767" stop-color="#FF4CE1" />
-        </radialGradient>
-        <linearGradient
-          id="paint3_linear_12184_111598"
-          x1="22.3999"
-          y1="51.0999"
-          x2="21"
-          y2="63.5"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop offset="0.085" stop-color="#FFBA33" />
-          <stop offset="0.553" stop-color="#FF006A" stop-opacity="0" />
-        </linearGradient>
-        <linearGradient
-          id="paint4_linear_12184_111598"
-          x1="21"
-          y1="51.5"
-          x2="21"
-          y2="63.5"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop offset="0.547" stop-color="white" stop-opacity="0" />
-          <stop offset="1" stop-color="white" stop-opacity="0.6" />
-        </linearGradient>
-        <linearGradient
-          id="paint5_linear_12184_111598"
-          x1="95"
-          y1="91"
-          x2="236.225"
-          y2="91"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop offset="0.0501219" stop-color="white" />
-          <stop offset="0.504634" stop-color="white" stop-opacity="0" />
-        </linearGradient>
-        <linearGradient
-          id="paint6_linear_12184_111598"
-          x1="64"
-          y1="91"
-          x2="263.5"
-          y2="91"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop offset="0.504634" stop-color="white" stop-opacity="0" />
-          <stop offset="0.85" stop-color="white" />
-        </linearGradient>
-      </defs>
-    </svg>
   );
 }
